split AsDefaultEloquentResource to two separate traits
